Create Scanner.c
Given the specifications for Pascal listed in the Attached to this Assignment, write a scanner that recognizes a function gettoken() that recognizes the lexemes of the language and returns a token corresponding to that lexeme.  

Since we do not yet have a symbol table, all words (a letter followed by zero or more alphanumeric characters) have the temporary token tokword, all numbers have the temporary token toknumber and all others have the temporary token tokop.